Wilderness Optimization
-Re enables the wilderness, with the temple and a rather poor crash
site.
-Updates the Lazy Mob spawner with some more optimizations, and a few
uses.
-Leaves a version of the crashsite with the original items and the new
system, I wasnt able to make it work without butchering it, and since I
am not that good of a mapper, I opted to leave it on the hands of people
that know what they are doing to improve the original, and making it
work, while leaving one that works.
-Changed the path of the teleports just to fuck up with Ketrai :)

Note to admins: Since this introduces 8 new z levels, all of which have
a check to generate or not a tree (think roughly a bit less than 100k
tiles checking for those), game init for objects went from 32 seconds to
71 seconds, so expect a rougher start on update.
Note to Ace: Does this counts as my Gateway Contest map? :P

additional notes: If performance takes a severe hit, init could be
severely reduced by reducing the trees chance to 2, or even placing
trees by hand on the maps.