Tweak out-of-ammo behavior on railguns
I personally like to imagine we live in a world where 'eject' is not a synonym for 'annihilate'.

Just makes the sound the railguns play when empty be a var, and makes them ACTUALLY eject the magazine instead of just deleting it (copypaste error probably, from the magnetic type, which does that).