Replaces NT Relation with Economic Status
Because starting money is pretty much all it affects, anyway. Instead of just being used to measure NT's opinion of you, you now choose income categories to fall under. This still affects how much money you start with, with the highest category giving a 1.3x modifier, but the lowest has been lowered to 0.5 times. Now you can start with even less money than before.

Currently doesn't change the "this person might be willing to help you" message for traitors, but it would be easy to do something different with that if requested.