Flooring fixes
Firstly it fixes a long-standing bugs with constructable flooring not actually populating on server startup, but instead through a proc that's only called on existing floor tiles.

Second it changes sif wood tiles into a child of wood tiles now that the reason for carpet colors not working is known & fixed.