Fixes a bug with integrated circuit printers. (#5407)
* Fixes a bug with integrated circuit printers.

This fixes a bug where tools, implants, and presumably (though i have yet to test it) device assemblies cannot be printed from the integrated circuit printer.
This is a much more robust solution for the print-anything bug than the previous one, as it expands the cost to the w_class of objects in it as well as lets you print anything in the current category. (This may be an issue when two people use it at once, but I mean, just learn to take turns or something.)

* closing parenthesis

oops