Adds Marking Order Changing at Character Creation (#5429)
* Adds Changing Order of Markings

Clicking the back button on a marking will trade its place with the previous marking or move it to the end of the list if it's the first marking.

Clicking the forward button on a marking will trade its place with the next marking or put it in first before all other markings if it's last in the list.

* Changelog

Adds changelog.

* Buttons Only for >1 Marking

Only renders the layer-swap buttons when there is more than one body marking.

* Changes Button Labels to Proper Up/Down Arrows

Thanks Atermonera!