Edit mapped in HoS armor
It was edited on the map to have pretty incredible armor values, better than SWAT armor, even. Things like this shouldn't be edited in the dmm files. Hard to have oversight on things like that.
Moved the armor values to the code and out of the map, and adjusted them to be more reasonable after discussing with other admins.