Fixes a runtime with radio.dm
Load ordering was causing the access datums for explorer and pilot access to be defined after radio.dm was loaded, which caused radio.dm to look for access datums that did not exist at the time.

This causes the explorer and pilot access datums to load much earlier, fixing the runtime.