Adds most of the remaining voremobs.
As with the previous PR this PR aims at adding most Voremobs outside of those with too much advanced mechanics that I simply can not port functionally.
Currently Oties and both Solargrub files are not present here.
The Rat file has a lot of gutted code but is otherwise functional as a non-hungry mob.