Protean Changes
- Changes Protean to Whitelist selectable
- Increases brute mod, decreases burn mod
- Halves Siemens coefficient. Zappy is still bad for Proteans, just not as bad.

I never was a fan of the 'Balance of extremes' of Proteans being near-immune to Brute and brute-related injuries, but a single shot with a laser or ion weapon practically vaporizes them.

However, Proteans are going to be problematic because they are basically Synths that get a ton of extra perks. They cannot get broken limbs, they can regenerate by just eating metal sheets, and they can get a huge number of bonuses by just eating metal sheets.

So, for that reason, some of those perks are probably going to be tweaked/reverted, and Proteans overall are going to take more brute/burn than other species in attempt to balance out those perks.