Changes the crew transfer message to be more IC friendly.
A simple change that changes the crew transfer vote message to something that implies the crew is voting for extra work on their PDAs.  This makes the crew transfer make a bit more sense in character, because as it stands, our characters are currently coming to work not really knowing when they get off work.

Direct port of this: https://github.com/Citadel-Station-13/Citadel-Station-13-RP/pull/572