Hopefully makes my fluff-stuff spawn WORN
Having a custom bag spawn in my internals box instead of on my back is annoying...also smol name edit because girlfren's character changed the last name spelling of their character and since our characters are married, gotta fix! 

but yeah hopefully this fixes it. it worked on private serber.