Prep For Revert of Removal
Couple minor changes with the substation name and maintenance: secondary command maint/substation now. The SMES, Powernet Sensor, and Breaker Box RCON tags are all renamed accordingly

This commit is for making the copy ready to use for copying over the old version of the Secondary Command, plus the maints and sand outside. The new file will be deleted for the second commit.