Allowing Custom Species access to DSI - Teshari prosthetic limbs.
Since Teshari are allowed to be created as a custom race to begin with, I'm not seeing too many issues with this change.

Tested on local and the main thing to note is that the user has to choose Icon Base as Teshari to avoid any weirdness with gear, otherwise works as expected.