Removes Healrod due to MISTAKES MADE
I managed to push the wrong sprite sets. This one's nearly impossible to see the charge meter of, and is just... Obnoxious.
Additionally, I've got new plans of how to make this darn thing, and thus I'm scrapping this project in favor of a new one! Thiiink I got all of them.