un-earrapifies instruments (upstream early access)
The default falloff value for sounds is 0.5 (lower means quieter).
There was only 3 things ingame with a sound falloff of a whole whopping 5 (10x)
These things were explosions, telepads, and instruments.
One had a volume of 25%, another one had a volume based on distance threshold, and you can probably guess which one of the bunch had a hardcoded 100% volume.