5 POI's (#7498)
* 5 POI's

As the tittle, adds 5 additional POI's intended for the plains.
Nothing valuable or useful should be found on any of these, maybe a bit of story and curiosity for anyone that wants to explore these.
Or better... maybe explore the remains and restore them to its former glory and take ownership to any abandoned home.

* Some fixes

-Changed name of 'abandonedhome.dmm' to 'lonehome.dmm'
-Changed 'lonehome.dmm' area, name and desc in their respective code so its all similar.
-Fixed 'priderock.dmm' once more as one will end in a endless tumbling loop. Should not happen anymore.

* Fix?

I changed all the id tags on electrochromatic windows and the button. See if this fixes Travis  being angy.