Other fixes
Hooked up the gas heating/cooling systems to the talon so they actually do something. Moved APC's to compensate and maintain symmetry.

Added Talon to the list of Radio channels that the portable radios on the Talon can communicate with.

Added armguards and leg-guards to the Talon Armory so that the space suit isn't the only thing that can cover arms and legs in a firefight.

Added a partslathe circutboard to the Engi-Vend on Talon.

Added Cameras to the Gravgen on Tether.

Added 30 sheets of plasteel to Talon's mat storage and a whetstone to Armory.