Remove broken, unused commcard system from communicators
This is being done to help facilitate the transfer to TGUI. These
commcards are part of a project that never really got finished, and the
majority of them do not even function in their current implementation,
with only barebone templates and half-made integration.

While it would be neat to have these function one day, I believe it
would be for the best to remove their current implementation wholesale
as this commit does, and allow them to be reintegrated at a later time
when communicators have been converted to TGUI and NanoUI has been
eliminated.