Spme more adjustments
So I want the Clown pet and the mime pet under c_pet.dm to change to an icon picked from a list of available icons when smacked. I have no error, but it just doesn't work.
Hinkle, the Clown Pet, can be found in the Clown office in TrashMaint, north of the tram.