MORE ALT ID STUFF
So basically I made one for all the station jobs, and then made one that when you use it, will become whatever job you are when you use it the first time. I added these to the loadout under utility as contractor identification cards!

I also did the same thing for the ITG IDs, but I did not put them on the loadout. People can fluff for them if they want to though! ITG is just a little more stringent in its hiring process than NT is.

I ALSO put in tons of pieces (which I used to make most of these) so that if someone wants to make more of them, or modularize some of them for whatever, they can! There's some alternative icon types and stuff in here too for making IDs for factions that don't exist or whatever! :0

I also removed Harmony's fluff ID and replaced it with the polymorphic ITG ID, since it's effectively the same thing but more generalized.