adds explorer equipment crate
adds a crate and fiddles with some points so that gearing up people we find and slow extra is possible from the vender. But much better done with the cargo crate. 10 cargo points for a crate was copied from the mining supply crate, I'd like it if someone would fiddle with the balance more.