Fix falling when moving diagonally past an open space.
This was happening because the decision to call handle_fall was made when over the open space, but execution deferred until the next tick.  By that time it is too late for the checks in handle_fall to abort falling, so lets call those in advance.
We do so in a separate proc from can_fall() because implementing it would be awkward, some types that override can_fall to avoid the anchored check would need to repeat the entire code or else delegate it.