Various map updates and things
Updates a paper's description to be more accurate after another change

corrects some tether related map errors on a few gateway maps

adds a pre-linked multitool for science for linking a Qpad to the science outpost. There's a pre-built Qpad on the science outpost as a result, but not one on the SD. You'll still have to build that yourself, but you won't HAVE to fly there. Mostly for in case there's no one to fly the ship around, it's still possible to get there and do work with a little effort, assuming there are materials for building a Qpad.