Removes the announcement for main pred and morph spawns
The announcement only really encourages people to go hunting for it, and usually not in the good way, which can in general ruin the surprise that these roles can employ.

So! I think it would be better if we just didn't announce them publicly, since they aren't a thing we ever want people to go and respond to mechanically, and, I think that announcements like these ought to be reserved for things we want people responding to.