Adds Temperature Adaptations to Xenochimera Trait List along with Autohiss.
Ran these by Scree (Creator of Xenochimera) and he said these are fairly reasonable adaptations.

They're neutral traits available to custom species, and given that Xenochimera are also able to pretty much mix and match a lot of stuff based on what they've eaten/evolved, this will allow greater roleplay/adaptations to occur - for instance, a xenochimera that is well-suited to the warmth of the desert is going to struggle in cold areas, and vice versa.

(It's conceivable that a Xenochimera might eat a wolf, and adapt/evolve a super-thick coat of fur suited for winter planets.)

Upstream port of https://github.com/CHOMPStation2/CHOMPStation2/pull/2948